Implemeting user roles

+++ app.py
@@ -2,10 +2,11 @@ from flask import Flask, request, jsonify
 from models.user import User
 from database import db
 from flask_login import LoginManager, login_user, current_user, logout_user, login_required
+import bcrypt
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = "your_secret_key"
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://root:admin123@127.0.0.1:3306/flask-crud'
 
 Login_manager = LoginManager()
 db.init_app(app)
@@ -17,6 +18,7 @@ Login_manager.login_view = 'login'
 def load_user(user_id):
   return User.query.get(user_id)
 
+# LOGIN DE USUARIO
 @app.route('/login', methods=["POST"])
 def login():
   data = request.json
@@ -27,48 +29,56 @@ def login():
     # login
     user = User.query.filter_by(username=username).first()
     
-    if user and user.password == password:
+    if user and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
       login_user(user)
       print(current_user.is_authenticated)
       return jsonify({"message": "Autenticação realizada com sucesso"})
   return jsonify({"message": "Credenciais inválidas"}), 400
 
+# LOGOUT DE USUARIO
 @app.route("/logout", methods=["GET"])
 @login_required
 def logout():
   logout_user()
   return jsonify({"message": "Logout realizado com sucesso!"})
 
+# CADASTRO DE USUARIO
 @app.route('/user', methods=["POST"])
-@login_required
 def create_user():
   data = request.json
   username = data.get("username")
   password = data.get("password")
 
   if username and password:
-      user = User(username=username, password=password)
-      db.session.add(user)
-      db.session.commit()
-      return jsonify({"message": "Usuario cadastrado com sucesso"})
+    hashed_password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
+    user = User(username=username, password=hashed_password, role='user')
+    db.session.add(user)
+    db.session.commit()
+    return jsonify({"message": "Usuario cadastrado com sucesso"})
 
   return jsonify({"message": "Dados invalidos"}), 400
 
+# BUSCAR USUARIO
 @app.route('/user/<int:id_user>', methods=["GET"])
 @login_required
 def read_user(id_user):
   user = User.query.get(id_user)
 
   if user:
-    return {"sername": user.username}
+    return {"username": user.username}
   
   return jsonify({"message": "Usuario não encontrado"}), 404
 
+# ATUALIZAR DADOS DE USUARIO
 @app.route('/user/<int:id_user>', methods=["PUT"])
 @login_required
 def update_user(id_user):
   data = request.json
   user = User.query.get(id_user)
+
+  if id_user != current_user.id and current_user.role == 'user':
+    return jsonify({"message": "Operação não permitida"}), 403
+
 
   if user and data.get("password"):
     user.password = data.get("password")
@@ -77,13 +87,17 @@ def update_user(id_user):
     return jsonify({"message": f"Usuário {id_user} atualizado com sucesso"})
   return jsonify({"message": f"Usuário não encontrado"}), 404
 
+# DELETAR USUARIO
 @app.route('/user/<int:id_user>', methods=["DELETE"])
 @login_required
 def delete_user(id_user):
   user = User.query.get(id_user)
 
+  if current_user.role != 'admin':
+    return jsonify({"message": "Operação não permitida"}), 403
+
   if id_user == current_user.id:
-    return jsonify({"message": f"Deleção não permitida"}), 403
+    return jsonify({"message": "Deleção não permitida"}), 403
 
   if user:
     db.session.delete(user)
